+++ src/Scripts/src/client/app.js
@@ -512,6 +512,22 @@
                          */
                         weavy.on("panel-close", bridgePanelEvent.bind(app, "close", panelId, { space: app.space, app: app }));
 
+                        /**
+                         * Triggered when the app receives a postMessage sent from the panel frame.
+                         * 
+                         * @category events
+                         * @event WeavyApp#message
+                         * @returns {Object}
+                         * @property {WeavySpace} space - The space that the app belongs to
+                         * @property {WeavyApp} app - The app that fires the event
+                         * @extends WeavyPanels#event:message
+                         */
+                        app.on("before:message", (e, message) => {
+                            if (message.panelId === panelId) {
+                                return WeavyUtils.assign(message, { space: app.space, app: app });
+                            }
+                        });
+
                         app.whenBuilt.resolve(app);
                     }
                 }
@@ -650,6 +666,25 @@
 
         return whenRemoved;
     }
+
+    /**
+     * Sends postMessage to the app panel frame. 
+     * Returns a promise that is resolved when the message has been delivered and rejected if the message fails or has timed out.
+     * 
+     * @category panel
+     * @function WeavyApp#postMessage
+     * @param {object} message - The Message to send
+     * @param {Transferable[]} [transfer] - A sequence of Transferable objects that are transferred with the message.
+     * @see {@link https://developer.mozilla.org/en-US/docs/Web/API/Window/postMessage}
+     * @returns {external:Promise}
+     * */
+    WeavyApp.prototype.postMessage = function (message, transfer) {
+        var app = this;
+        return app.whenBuilt().then(function () {
+            return app.panel.postMessage(message, transfer);
+        });
+    }
+
 
     /**
      * Check if another app or an object is matching this app. It checks for a match of the id property or the key property.